Isolate and reuse subscribeToOneSearch() helpers
